refactor: ServiceTest 제거 및 GameTest 추가
